feat(command-blur): Remove click listener in favor of CSS pointer-events
When workbench.quickOpen.closeOnFocusLost is set to false, clicking
elsewhere on the screen closes the #command-blur div while the command
palette remains open.

This change makes the opening and closing of the #command-blur div
respect the value set in workbench.quickOpen.closeOnFocusLost, closing
it according to the state of the command palette correctly.